fix: NeoForge mixin error
