+++ common/src/main/java/com/daqem/grieflogger/event/block/BlockEvents.java
@@ -1,5 +1,6 @@
 package com.daqem.grieflogger.event.block;
 
+import dev.architectury.event.EventResult;
 import dev.architectury.event.events.common.BlockEvent;
 import dev.architectury.event.events.common.InteractionEvent;
 
@@ -7,7 +8,10 @@ public class BlockEvents {
 
     public static void registerEvents() {
         BlockEvent.BREAK.register(BreakBlockEvent::breakBlock);
-        BlockEvent.PLACE.register(PlaceBlockEvent::placeBlock);
+        BlockEvent.PLACE.register(((level, blockPos, blockState, entity) -> {
+            PlaceBlockEvent.placeBlock(level, blockPos, blockState, entity);
+            return EventResult.pass();
+        }));
         InteractionEvent.LEFT_CLICK_BLOCK.register((player, hand, pos, direction) -> LeftClickBlockEvent.leftClickBlock(player, hand, pos, direction).asMinecraft());
         InteractionEvent.RIGHT_CLICK_BLOCK.register((player, hand, pos, direction) -> RightClickBlockEvent.rightClickBlock(player, hand, pos, direction).asMinecraft());
     }
