+++ common/src/main/java/com/daqem/grieflogger/mixin/MixinBucketItem.java
@@ -23,20 +23,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class MixinBucketItem {
 
     @Inject(
-            method = "emptyContents",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/world/level/Level;setBlock(Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;I)Z",
-                    shift = At.Shift.AFTER
-            )
-    )
-    private void onLiquidPlaced(LivingEntity livingEntity, Level level, BlockPos blockPos, BlockHitResult blockHitResult, CallbackInfoReturnable<Boolean> cir) {
-        if (livingEntity instanceof ServerPlayer serverPlayer) {
-            PlaceBlockEvent.placeBlock(level, blockPos, level.getBlockState(blockPos), serverPlayer);
-        }
-    }
-
-    @Inject(
             method = "use",
             at = @At(
                     value = "INVOKE",
