+++ fabric/src/main/java/com/daqem/grieflogger/fabric/mixin/MixinBucketItem.java
@@ -1,0 +1,31 @@
+package com.daqem.grieflogger.fabric.mixin;
+
+import com.daqem.grieflogger.event.block.PlaceBlockEvent;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.BucketItem;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.BlockHitResult;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(BucketItem.class)
+public class MixinBucketItem {
+
+    @Inject(
+            method = "emptyContents",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/world/level/Level;setBlock(Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;I)Z",
+                    shift = At.Shift.AFTER
+            )
+    )
+    private void onLiquidPlaced(LivingEntity livingEntity, Level level, BlockPos blockPos, BlockHitResult blockHitResult, CallbackInfoReturnable<Boolean> cir) {
+        if (livingEntity instanceof ServerPlayer serverPlayer) {
+            PlaceBlockEvent.placeBlock(level, blockPos, level.getBlockState(blockPos), serverPlayer);
+        }
+    }
+}
