+++ gradle.properties
@@ -5,7 +5,7 @@ maven_group=com.daqem
 archives_base_name=grieflogger
 
 # Project
-mod_version=1.2.7-1.21.6
+mod_version=1.2.8-1.21.6
 mod_id=grieflogger
 mod_name=GriefLogger
 mod_description=A mod that logs all player interactions with blocks and entities and stores them in a database.
