+++ neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -42,3 +42,6 @@ side = "SERVER"
 
 [[mixins]]
 config = "grieflogger-common.mixins.json"
+
+[[mixins]]
+config = "grieflogger-neoforge.mixins.json"
