+++ settings.gradle
@@ -11,4 +11,4 @@ include 'common'
 include 'fabric'
 include 'neoforge'
 
-rootProject.name = "GriefLogger 1.21.5"
+rootProject.name = "GriefLogger 1.21.6"
